# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 経理・会計オールインワンAIアプリケーション API
  version: "1.0.0"
  description: |
    中小企業〜中堅企業向け「経理・会計をこのアプリ1つで完結させる」オールインワンAIプロダクトのAPI仕様。

    対応する統合DB設計は `docs/03_database_design.md` / `sql/001_initial_schema_all_in_one.sql` を参照。

    ## 設計原則
    - AIは提案(`/ai/suggestions`)のみを行い、確定処理(仕訳確定・税額計算等)は本APIの決定的ロジックが行う。
    - 状態遷移・業務アクションはサブリソース(例: `POST /journal-entries/{id}/post`)として明示する。
    - 全エンドポイントはマルチテナントであり、`X-Tenant-ID` ヘッダーとJWTクレームの双方で検証される。
    - 副作用を伴うPOST操作は `Idempotency-Key` ヘッダーによる冪等性をサポートする。
  contact:
    name: 経理会計オールインワンソフト開発チーム
servers:
  - url: https://api.keiri-kaikei.example.com/v1
    description: 本番環境
  - url: https://api.staging.keiri-kaikei.example.com/v1
    description: ステージング環境
  - url: http://localhost:3000/v1
    description: ローカル開発環境
...
security:
  - BearerAuth: []

# ==============================================================================
# PATHS
# ==============================================================================
paths:
...
  /invoices:
    get:
      tags: [Invoices]
      summary: 売上請求書を検索する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/InvoiceStatus" }
        - name: customer_id
          in: query
          schema: { type: string, format: uuid }
        - name: due_date_from
          in: query
          schema: { type: string, format: date }
        - name: due_date_to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: 売上請求書一覧
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceListResponse" }
    post:
      tags: [Invoices]
      summary: 売上請求書を草案(draft)として作成する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvoiceCreate" }
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceResponse" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /invoices/{id}:
    get:
      tags: [Invoices]
      summary: 売上請求書詳細(明細含む)を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: 売上請求書詳細
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Invoices]
      summary: 売上請求書を編集する(draft状態のみ)
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvoiceCreate" }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceResponse" }
        "409": { $ref: "#/components/responses/Conflict" }
  /invoices/{id}/issue:
    post:
      tags: [Invoices]
      summary: 売上請求書を発行する(売上計上仕訳を自動起票し、以後は追記専用化)
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      responses:
        "200":
          description: 発行成功(status=issued)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceResponse" }
        "409": { $ref: "#/components/responses/Conflict" }
  /invoices/{id}/void:
    post:
      tags: [Invoices]
      summary: 売上請求書を取消す(発行後24時間以内かつ未入金の場合のみ)
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      responses:
        "200":
          description: 取消成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoiceResponse" }
        "409": { $ref: "#/components/responses/Conflict" }
  /invoices/{id}/credit-notes:
    post:
      tags: [Invoices]
      summary: 発行済み請求書に対する訂正(クレジットノート)を起票する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, reason]
              properties:
                amount: { type: number }
                reason: { type: string }
      responses:
        "201":
          description: クレジットノート作成成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreditNoteResponse" }
  /invoices/{id}/payments:
    get:
      tags: [Invoices]
      summary: 請求書に対する入金消込履歴を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: 入金消込一覧
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoicePaymentListResponse" }
    post:
      tags: [Invoices]
      summary: 売掛金の入金消込を手動で記録する(一部入金・過入金に対応)
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payment_date, amount]
              properties:
                payment_date: { type: string, format: date }
                amount: { type: number }
                bank_transaction_id: { type: string, format: uuid }
      responses:
        "201":
          description: |
            消込成功。全額消込の場合 invoice.status=paid、
            一部消込の場合は partially_paid へ遷移する。
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvoicePaymentResponse" }

  # ----------------------------------------------------------------------------
  # Vendor Bills (仕入請求書/買掛金 = Accounts Payable)
  # ----------------------------------------------------------------------------
...
  /payment-batches:
    get:
      tags: [PaymentBatches]
      summary: 支払バッチ一覧を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/PageParam"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/PaymentBatchStatus" }
      responses:
        "200":
          description: 支払バッチ一覧
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentBatchListResponse" }
  /payment-batches/{id}:
    get:
      tags: [PaymentBatches]
      summary: 支払バッチ詳細(明細含む)を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: 支払バッチ詳細
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentBatchResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /payment-batches/export-zengin:
    post:
      tags: [PaymentBatches]
      summary: 承認済みの支払対象(仕入請求書/経費立替金/給与)を集計し全銀協FBデータを生成する
      description: |
        対象明細はこの時点でスナップショットとして `payment_batch_items` に固定される
        (以後の請求書変更の影響を受けない)。生成後は `status=exported` となり、
        再生成不可(改版時は新バッチを作成する)。
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payment_date, sources]
              properties:
                payment_date: { type: string, format: date }
                sources:
                  type: array
                  description: 支払対象の発生源を横断的に指定する(未指定項目は対象外)
                  items:
                    type: object
                    required: [source_type, source_id]
                    properties:
                      source_type:
                        type: string
                        enum: [vendor_bill, expense_reimbursement, payroll]
                      source_id: { type: string, format: uuid }
      responses:
        "201":
          description: 全銀協FBデータ生成成功(ダウンロードURLを含む)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaymentBatchResponse"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PaymentBatch"
                          - type: object
                            properties:
                              download_url:
                                type: string
                                format: uri
        "422": { $ref: "#/components/responses/UnprocessableEntity" }

  # ----------------------------------------------------------------------------
  # Expense Reports (領域B)
  # ----------------------------------------------------------------------------
...
  /payroll-imports:
    get:
      tags: [PayrollImports]
      summary: 給与インポート履歴一覧を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/PageParam"
      responses:
        "200":
          description: 給与インポート履歴一覧
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayrollImportListResponse" }
  /payroll-imports/{id}:
    get:
      tags: [PayrollImports]
      summary: 給与インポート詳細(従業員別明細含む)を取得する
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: 給与インポート詳細
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayrollImportResponse" }
        "404": { $ref: "#/components/responses/NotFound" }
  /payroll-imports/csv:
    post:
      tags: [PayrollImports]
      summary: |
        給与ソフトCSVを取込み、役員報酬/給与手当/預り金各種/法定福利費への
        複合仕訳(draft)を1トランザクションで自動生成する。
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, import_mapping_id, pay_period_start, pay_period_end, payment_date]
              properties:
                file: { type: string, format: binary }
                import_mapping_id: { type: string, format: uuid }
                pay_period_start: { type: string, format: date }
                pay_period_end: { type: string, format: date }
                payment_date: { type: string, format: date }
      responses:
        "201":
          description: 取込・複合仕訳生成成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayrollImportResponse" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /payroll-imports/{id}/post:
    post:
      tags: [PayrollImports]
      summary: 複合仕訳を確定する(accounting_managerによる最終承認が必須)
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdPathParam"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      responses:
        "200":
          description: 確定成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PayrollImportResponse" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # ----------------------------------------------------------------------------
  # AI Suggestions (AI提案の隔離)
  # ----------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWTアクセストークン。`tenant_id` クレームを含み、`X-Tenant-ID` ヘッダーと
        一致しない場合はAPI層で403 (`TENANT_MISMATCH`) を返す。
  parameters:
    TenantIdHeader:
      name: X-Tenant-ID
      in: header
      required: true
      description: 操作対象テナントのUUID。JWTクレームの tenant_id と一致する必要がある。
      schema: { type: string, format: uuid }
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        副作用を伴うPOST操作の冪等性を保証するクライアント生成キー(推奨: UUID)。
        同一キーでの再送は最初のレスポンスをそのまま返す。
      schema: { type: string }
    IdPathParam:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
...
    Forbidden:
      description: 権限がありません(RBAC不足、またはX-Tenant-IDとJWTクレームの不一致)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: リソースが見つかりません(RLSにより越境アクセス時も同一の404を返す)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: 状態遷移ルール・DB制約違反等によりリクエストを実行できません
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    UnprocessableEntity:
      description: 入力値のバリデーションエラー
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    Meta:
      type: object
      properties:
        request_id: { type: string, format: uuid }
        pagination:
          type: object
          properties:
            page: { type: integer }
            page_size: { type: integer }
            total_count: { type: integer }
            total_pages: { type: integer }
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            例: INVALID_BALANCE, VALIDATION_ERROR, NOT_FOUND, UNAUTHORIZED, FORBIDDEN,
            TENANT_MISMATCH, IDEMPOTENCY_KEY_CONFLICT, VOID_WINDOW_EXPIRED,
            SELF_APPROVAL_NOT_ALLOWED, DUPLICATE_IMPORT, CONFLICT, ALREADY_EXISTS,
            INVITATION_EXPIRED, INVITATION_REVOKED, INVITATION_ALREADY_ACCEPTED
        message: { type: string }
        details:
          type: array
          items: { type: object }
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { const: false }
        error: { $ref: "#/components/schemas/ApiError" }

    # ---- Enums (DBのENUM型と対応) ----
...
    InvoiceStatus:
      type: string
      enum: [draft, issued, partially_paid, paid, overdue, voided, credit_note_issued]
...
    PaymentBatchStatus:
      type: string
      enum: [draft, exported, completed, cancelled]
...
    InvoiceLine:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        line_no: { type: integer }
        description: { type: string }
        quantity: { type: number, default: 1 }
        unit_price: { type: number }
        amount: { type: number }
        tax_category_id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
    InvoiceLineCreate:
      type: object
      required: [description, unit_price, tax_category_id, account_id]
      properties:
        description: { type: string }
        quantity: { type: number, default: 1 }
        unit_price: { type: number }
        tax_category_id: { type: string, format: uuid }
        account_id: { type: string, format: uuid }
    Invoice:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        invoice_no: { type: string, readOnly: true }
        customer_id: { type: string, format: uuid }
        issue_date: { type: string, format: date }
        due_date: { type: string, format: date }
        status: { $ref: "#/components/schemas/InvoiceStatus" }
        subtotal_amount: { type: number, readOnly: true }
        tax_amount: { type: number, readOnly: true }
        total_amount: { type: number, readOnly: true }
        currency_code: { type: string, default: JPY }
        journal_entry_id: { type: string, format: uuid, nullable: true, readOnly: true }
        issued_at: { type: string, format: date-time, nullable: true, readOnly: true }
        lines:
          type: array
          items: { $ref: "#/components/schemas/InvoiceLine" }
    InvoiceCreate:
      type: object
      required: [customer_id, issue_date, due_date, lines]
      properties:
        customer_id: { type: string, format: uuid }
        issue_date: { type: string, format: date }
        due_date: { type: string, format: date }
        currency_code: { type: string, default: JPY }
        lines:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/InvoiceLineCreate" }
    InvoiceResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/Invoice" }
        meta: { $ref: "#/components/schemas/Meta" }
    InvoiceListResponse:
      type: object
      properties:
        success: { const: true }
        data:
          type: array
          items: { $ref: "#/components/schemas/Invoice" }
        meta: { $ref: "#/components/schemas/Meta" }
    CreditNote:
      type: object
      readOnly: true
      properties:
        id: { type: string, format: uuid }
        original_invoice_id: { type: string, format: uuid }
        credit_note_no: { type: string }
        issue_date: { type: string, format: date }
        amount: { type: number }
        reason: { type: string }
        journal_entry_id: { type: string, format: uuid, nullable: true }
    CreditNoteResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/CreditNote" }
        meta: { $ref: "#/components/schemas/Meta" }
    InvoicePayment:
      type: object
      readOnly: true
      properties:
        id: { type: string, format: uuid }
        invoice_id: { type: string, format: uuid }
        payment_date: { type: string, format: date }
        amount: { type: number }
        bank_transaction_id: { type: string, format: uuid, nullable: true }
        journal_entry_id: { type: string, format: uuid, nullable: true }
        matched_by: { type: string, enum: [auto, manual] }
    InvoicePaymentResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/InvoicePayment" }
        meta: { $ref: "#/components/schemas/Meta" }
    InvoicePaymentListResponse:
      type: object
      properties:
        success: { const: true }
        data:
          type: array
          items: { $ref: "#/components/schemas/InvoicePayment" }
        meta: { $ref: "#/components/schemas/Meta" }

    # ---- Vendor Bills (AP) ----
...
    PaymentBatchItem:
      type: object
      readOnly: true
      properties:
        id: { type: string, format: uuid }
        source_type: { type: string, enum: [vendor_bill, expense_reimbursement, payroll] }
        source_id: { type: string, format: uuid }
        payee_name: { type: string }
        amount: { type: number }
    PaymentBatch:
      type: object
      readOnly: true
      properties:
        id: { type: string, format: uuid }
        batch_no: { type: string }
        payment_date: { type: string, format: date }
        batch_type: { type: string, enum: [zengin_transfer] }
        status: { $ref: "#/components/schemas/PaymentBatchStatus" }
        total_amount: { type: number }
        file_hash: { type: string, nullable: true }
        exported_at: { type: string, format: date-time, nullable: true }
        items:
          type: array
          items: { $ref: "#/components/schemas/PaymentBatchItem" }
    PaymentBatchResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/PaymentBatch" }
        meta: { $ref: "#/components/schemas/Meta" }
    PaymentBatchListResponse:
      type: object
      properties:
        success: { const: true }
        data:
          type: array
          items: { $ref: "#/components/schemas/PaymentBatch" }
        meta: { $ref: "#/components/schemas/Meta" }

    # ---- Expense Reports ----
...
    PayrollImportLine:
      type: object
      readOnly: true
      properties:
        id: { type: string, format: uuid }
        employee_name: { type: string }
        employee_code: { type: string, nullable: true }
        executive_compensation_amount: { type: number }
        salary_amount: { type: number }
        withholding_tax_amount: { type: number }
        resident_tax_amount: { type: number }
        social_insurance_employee_amount: { type: number }
        social_insurance_company_amount: { type: number }
        net_payment_amount: { type: number }
    PayrollImport:
      type: object
      readOnly: true
      properties:
        id: { type: string, format: uuid }
        pay_period_start: { type: string, format: date }
        pay_period_end: { type: string, format: date }
        payment_date: { type: string, format: date }
        status: { type: string, enum: [imported, posted] }
        journal_entry_id: { type: string, format: uuid, nullable: true }
        lines:
          type: array
          items: { $ref: "#/components/schemas/PayrollImportLine" }
    PayrollImportResponse:
      type: object
      properties:
        success: { const: true }
        data: { $ref: "#/components/schemas/PayrollImport" }
        meta: { $ref: "#/components/schemas/Meta" }
    PayrollImportListResponse:
      type: object
      properties:
        success: { const: true }
        data:
          type: array
          items: { $ref: "#/components/schemas/PayrollImport" }
        meta: { $ref: "#/components/schemas/Meta" }

    # ---- Payroll Import Mappings ----
...